Struct sysly::Syslog

pub struct Syslog {
    // some fields omitted
}

A rust interface for Syslog, a standard unix system logging service

Methods

impl Syslog

fn udp(host: SocketAddr) -> Syslog

Factory for a Syslog appender that writes to remote Syslog daemon listening a SocketAddr

fn localudp() -> Syslog

Same as udp with providing local loopback address with the standard syslog port

fn unix<P: AsRef<Path>>(path: P) -> Syslog

Factory for a Syslog appender that writes to a host-local Syslog daemon listening on a unix socket domain hosted at the given Path

fn facility(self, facility: Facility) -> Syslog

Returns a new Syslog appender configured to append with the provided Facility

fn host(self, local: &str) -> Syslog

Returns a new Syslog appender configured to append with the provided host addr

fn app(self, app: &str) -> Syslog

Returns a new Syslog appender, configured to append with the provided app-name

fn pid(self, pid: &str) -> Syslog

Returns a new Syslog appender configured to append with the provided p(rocess)id

fn msgid(self, id: &str) -> Syslog

Returns a new Syslog appender configured to append with the provided msgid

fn debug(&mut self, msg: &str) -> Result

Emits a debug level message

fn info(&mut self, msg: &str) -> Result

Emits an info level message

fn notice(&mut self, msg: &str) -> Result

Emits an info level message

fn warn(&mut self, msg: &str) -> Result

Emits an warn level message

fn err(&mut self, msg: &str) -> Result

Emits an error level message

fn critical(&mut self, msg: &str) -> Result

Emits a critical level message

fn alert(&mut self, msg: &str) -> Result

Emits an alert level message

fn emergency(&mut self, msg: &str) -> Result

Emits a emergencycritical level message